backend/requirements.txt
accelerate==1.4.0
annotated-types==0.7.0
anyio==4.8.0
av==14.2.0
bitsandbytes==0.45.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
easyocr==1.7.2
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
fastapi-cli==0.0.7
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
intel-openmp==2021.4.0
Jinja2==3.1.5
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.3
numpy==1.26.3
onnx==1.17.0
onnxruntime==1.19.2
onnxruntime-gpu==1.19.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==6.30.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pyreadline3==3.5.4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
qwen-vl-utils==0.0.10
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
safetensors==0.5.3
scikit-image==0.24.0
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
sympy==1.13.1
tbb==2021.13.1
tifffile==2024.8.30
tokenizers==0.21.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
transformers==4.49.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.89
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0
xxhash==3.5.0
zipp==3.21.0
zstandard==0.23.0
